Python-Projects/MNIST-MultiModel-FastAPI-API/requirements.txt
# Python-Projects/MNIST-MultiModel-FastAPI-API/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pillow==10.1.0
opencv-python-headless==4.8.1.78
transformers==4.35.2
pydantic==2.5.0
python-multipart==0.0.6
